+++ changelog.md
@@ -4,3 +4,8 @@
 - added `environment.yml`
 - re-arranged code into smaller modules
 - added test step to be used for final quality checks after training (loss, R2, relative error, clustering metric)
+
+
+[SK] - 28.04.2023:
+- added reconstruction accuracy metrices for scoring (table in tensorboard)
+- added embedding view in tensorboard